# backend/swagger.yaml
openapi: 3.1.0
info:
  title: Mixillo API
  version: 1.0.0
  description: |
    Complete API documentation for Mixillo - TikTok-style social commerce platform.
    
    ## Features
    - 🔐 JWT Authentication with refresh tokens
    - 📹 TikTok-style video content & social feeds
    - 🛍️ E-commerce with cart & order management
    - 💰 Digital wallet & coin system
    - 🎁 Virtual gifting & monetization
    - 📺 Live streaming with Agora SDK
    - 👥 Social features (follow, like, comment, share)
    - 📊 Analytics & trending algorithms
    - 🛡️ AI-powered content moderation
    - 📱 Multi-language & multi-currency support
    
  contact:
    name: Mixillo Support
    email: support@mixillo.com
  license:
    name: Proprietary
servers:
  - url: https://mixillo-backend-52242135857.europe-west1.run.app
    description: Production Server (Google Cloud Run)
  - url: http://localhost:5000
    description: Local Development Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
...
paths:
...
  /api/admin/analytics:
    get:
      tags:
        - Admin
      summary: Get system analytics (admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: System analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      totalUsers:
                        type: integer
                        example: 10523
                      totalRevenue:
                        type: number
                        example: 245678.50
                      totalOrders:
                        type: integer
                        example: 3456
                      totalContent:
                        type: integer
                        example: 45678
                      activeStreams:
                        type: integer
                        example: 12
        '403':
          description: Admin access required
  
  # ==================== SYSTEM ====================
...